move pubsub tasks to core
