+++ server/src/party-bookings/core/send-party-form-reminder-emails.ts
@@ -2,10 +2,10 @@ import { logger } from 'firebase-functions/v2'
 import type { Booking } from 'fizz-kidz'
 import { getManager } from 'fizz-kidz'
 
-import { FirestoreRefs } from '../../../firebase/FirestoreRefs'
-import { MailClient } from '../../../sendgrid/MailClient'
-import { logError } from '../../../utilities'
-import { getPrefilledFormUrl, getUpcoming } from '../../core/utils.party'
+import { FirestoreRefs } from '../../firebase/FirestoreRefs'
+import { MailClient } from '../../sendgrid/MailClient'
+import { logError } from '../../utilities'
+import { getPrefilledFormUrl, getUpcoming } from './utils.party'
 
 export async function sendPartyFormReminderEmails() {
     // since this runs on a Monday, it will get tomorrow
