+++ server/src/pubsub.ts
@@ -1,10 +1,10 @@
 import { onMessagePublished, logError } from './utilities'
-import { sendIncursionForms } from './events/functions/pubsub/send-incursion-forms'
-import { sendGuestsEmail } from './party-bookings/functions/pubsub/send-guests-email'
-import { sendPartyFormReminderEmails } from './party-bookings/functions/pubsub/send-party-form-reminder-emails'
-import { sendPartyForms } from './party-bookings/functions/pubsub/send-party-forms'
-import { sendPartyFeedbackEmails } from './party-bookings/functions/pubsub/send-party-feedback-emails'
-import { remindAboutWwcc } from './staff/functions/pubsub/remind-about-wwcc'
+import { sendIncursionForms } from './events/core/send-incursion-forms'
+import { sendGuestsEmail } from './party-bookings/core/send-guests-email'
+import { sendPartyFormReminderEmails } from './party-bookings/core/send-party-form-reminder-emails'
+import { sendPartyForms } from './party-bookings/core/send-party-forms'
+import { sendPartyFeedbackEmails } from './party-bookings/core/send-party-feedback-emails'
+import { remindAboutWwcc } from './staff/core/remind-about-wwcc'
 import { handlePaperformSubmission } from './paperforms/functions/pubsub/paperform.pubsub'
 import type { PubSubFunctions } from 'fizz-kidz'
